Add new GenShell to Fakes Test project
